STARTREK: Move space functions into separate file
